# apps/weasq-bpt-dashboard/bpt-api.spec.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "https://api.weasq.io/bpt", "description": "Generated server url" }],
...
  "paths": {
...
    "/tags/{id}": {
      "get": {
        "tags": ["Tag"],
        "summary": "Get Tag by id",
        "operationId": "getTagData",
        "parameters": [
          { "name": "Authorization", "in": "header", "required": true, "schema": { "type": "string" } },
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be updated",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          },
          "404": {
            "description": "Tag not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          }
        }
      },
      "put": {
        "tags": ["Tag"],
        "summary": "Updated Tag",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateTag",
        "parameters": [
          { "name": "Authorization", "in": "header", "required": true, "schema": { "type": "string" } },
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be updated",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagDataResource" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          },
          "400": {
            "description": "Invalid Tag supplied",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          },
          "404": {
            "description": "Tag not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          }
        }
      },
      "delete": {
        "tags": ["Tag"],
        "summary": "Delete Tag",
        "description": "This can only be done by the logged in User.",
        "operationId": "deleteTag",
        "parameters": [
          { "name": "Authorization", "in": "header", "required": true, "schema": { "type": "string" } },
          {
            "name": "id",
            "in": "path",
            "description": "id that need to be updated",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "201": { "description": "operation successful" },
          "400": { "description": "Invalid id supplied" },
          "404": { "description": "Umfrage not found" }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["Tag"],
        "summary": "Get all Tags",
        "operationId": "getTagList",
        "parameters": [{ "name": "Authorization", "in": "header", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TagResource" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tag"],
        "summary": "Create new Tag",
        "description": "This can only be done by the logged in user",
        "operationId": "postTag",
        "parameters": [{ "name": "Authorization", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagDataResource" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          },
          "400": {
            "description": "Invalid Tag supplied",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          },
          "404": {
            "description": "Tag not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagResource" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TagResource": {
        "type": "object",
        "properties": { "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" } }
      },
...
      "TagDataResource": {
        "type": "object",
        "properties": { "name": { "type": "string" } },
        "description": "Tag to add"
      },
...
    }
  }
}